Extract reporting query contracts
